Refactor virtual user handler
